1-Python/rpa/requirements.txt
imbox==0.9.8
openpyxl==3.1.2
pdfplumber==0.10.3
pillow==10.2.0
pypdf==3.17.4